Partial support for AutoScaling RayCluster
For a RayCluster object with spec.enableInTreeAutoscaling=True there is one
direct PodSet object associated with it which involves the head node pod.

Ray is then allowed to create Pods which all come with a "kueue.x-k8s.io/queue-name"
label pointing to a LocalQueue. Thereby, enabling Kueue to create Workload objects
for these Pods.

When auto scaling is switched off, there's a single Workload object for all the
pods associated with the RayCluster object.

The current approach is not working as intended when auto scaling is enabling:
- Ray creates pods for its autoscaling WorkerGroupSpec definitions
- Those pods have a kueue.x-k8s.io/queue-name label
- However, the pods do not receive the kueue.x-k8s.io/managed label. Also,
  their resources do not count against the resource quota of the LocalQueue

Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>